server/pyproject.toml
[project]
name = "english-article-interpretation-server"
version = "0.1.0"
description = "Claread透读后端服务"
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    "fastapi>=0.116.1,<0.117.0",
    "httpx>=0.28.1,<0.29.0",
    "langgraph>=0.6.6,<0.7.0",
    "langsmith[otel]>=0.4.26,<0.5.0",
    "opentelemetry-exporter-otlp>=1.39.0,<2.0.0",
    "orjson>=3.11.1,<4.0.0",
    "pydantic-ai>=1.0.14,<2.0.0",
    "pydantic-settings>=2.10.1,<3.0.0",
    "uvicorn[standard]>=0.35.0,<0.36.0",
    # 数据库
    "asyncpg>=0.30.0,<0.31.0",
    "beautifulsoup4>=4.12.3,<5.0.0",
    # Redis（可选，redis_enabled=False 时不阻塞启动）
    "redis>=5.2.0,<6.0.0",
]

[dependency-groups]
dev = [
    "mypy>=1.17.1,<2.0.0",
    "pytest>=8.4.1,<9.0.0",
    "pytest-asyncio>=0.24.0,<0.25.0",
    "ruff>=0.12.8,<0.13.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_configs = true
plugins = ["pydantic.mypy"]
